plotando os peões brancos

+++ src/scene/background.c
@@ -11,8 +11,8 @@ void setup_background(int w, int h){
 
     image_board = LoadTexture("../assets/board.png"); 
 
-    recSrc = (Rectangle){0, 0, image_board.width, image_board.height};
-    recDest = (Rectangle){0, 0, w, h};
+    recSrc = (Rectangle){0, 0, (float)image_board.width, (float)image_board.height};
+    recDest = (Rectangle){0, 0, (float)w, (float)h};
     origin = (Vector2){0, 0};
 }
